requirements.txt
amqp==2.6.1
asgiref==3.6.0
billiard==3.6.4.0
cached-property==1.5.2
celery==4.4.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.18
djangorestframework==3.14.0
greenlet==2.0.2
importlib-metadata==4.13.0
kombu==4.6.11
prompt-toolkit==3.0.38
pytz==2022.7.1
six==1.16.0
SQLAlchemy==1.4.0
sqlparse==0.4.3
typing_extensions==4.5.0
vine==1.3.0
wcwidth==0.2.6
zipp==3.15.0
